+++ CHANGELOG.md
@@ -7,7 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.1] 2020-05-22
+## [1.0.4] 2020-05-22
+
+### Fixed
+
+- Avoid exporting metrics for deleted clusters.
+
+## [1.0.3] 2020-05-22
+
+### Fixed
+
+- Fixed prometheus exporter.
+
+## [1.0.2] 2020-05-22
 
 ### Fixed
 
@@ -25,7 +37,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/etcd-backup-operator/compare/v1.0.2...HEAD
+
+[Unreleased]: https://github.com/giantswarm/etcd-backup-operator/compare/v1.0.4...HEAD
+[1.0.4]: https://github.com/giantswarm/etcd-backup-operator/compare/v1.0.3...v1.0.4
+[1.0.3]: https://github.com/giantswarm/etcd-backup-operator/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/giantswarm/etcd-backup-operator/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/giantswarm/etcd-backup-operator/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/giantswarm/etcd-backup-operator/releases/tag/v1.0.0